requirements/dev.txt
-r base.txt
django-debug-toolbar
django-extensions
werkzeug